# pangea-api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Category:
    properties:
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_featured:
        type: boolean
      name:
        type: string
      parent_id:
        type: integer
      slug:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    required:
    - name
    type: object
  models.CategoryRequest:
    properties:
      color:
        type: string
      description:
        type: string
      icon:
        type: string
      is_active:
        type: boolean
      is_featured:
        type: boolean
      name:
        type: string
      parent_id:
        type: integer
      sort_order:
        type: integer
    required:
    - name
    type: object
  models.CategoryResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      category:
        $ref: '#/definitions/models.Category'
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Pangea UI Backend API with authentication, admin panel, and OAuth integration
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pangea API
  version: "1.0"
paths:
...
  /api/v1/categories:
    get:
      consumes:
      - application/json
      description: Retrieve all categories
      parameters:
      - description: Filter by active status
        in: query
        name: active
        type: boolean
      - description: Filter by featured status
        in: query
        name: featured
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CategoryResponse'
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new category
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CategoryResponse'
      summary: Create new category
      tags:
      - categories
  /api/v1/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.CategoryResponse'
      summary: Delete category
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Retrieve a specific category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.CategoryResponse'
      summary: Get category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update an existing category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.CategoryResponse'
      summary: Update category
      tags:
      - categories
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"